Changed structure on how hubs will send and receive data +++Added two new members to the radio class: requestData & waitForRequest >>>Both functions are a part of the new system ===New system: >>>'slave' hubs will only send data if anyother hub requests it >>>right now only the master hub will request data, but in the future other hubs will be able to request data and their requests will be put in FIFO (queue) >>>this helps prevent multiple 'slave' hubs from contacting the master at once ===Have tested the code and works perfectly >>>hub 9002 will sends data only when 9001 requests it

+++ SensorHub/constants.h
@@ -8,15 +8,17 @@
 //																							    //
 //**********************************************************************************************//
 #pragma once
+#include "Arduino.h"
+
 const int STATION[8] = { 9001, 9002, 9003, 9004, 9005, 9006, 9007, 9008 }; //Set the addresses for all stations
 
 //Set Arduino pins
-const int SWITCH_IN_1 = 2;
-const int SWITCH_IN_2 = 3;
-const int SWITCH_IN_3 = 4;
-const int ERROR_LED = 7;
-const int PIR = 5;
-const int DEBUG_LED = 8;
+const byte SWITCH_IN_1 = 2;
+const byte SWITCH_IN_2 = 3;
+const byte SWITCH_IN_3 = 4;
+const byte ERROR_LED = 7;
+const byte PIR = 5;
+const byte DEBUG_LED = 8;
 
 //Timer consts
 const int SEND_TIME = 10000; //Time it takes a hub to completly send a message
